Relocate package prefix to org.scijava.vecmath
This avoids a serious problem with Java 3D on OS X 10.11 "El Capitan":
it is no longer possible for users to easily delete the Java 3D 1.3
artifacts out of /System/Library/Java/Extensions, thanks to El Capitan's
new System Integrity Protection (SIP, a.k.a "rootless") feature.

So, even with the new Java 3D 1.6 on the classpath, the old Java 3D 1.3
always takes precedence because /System/Library/Java/Extensions is on
OS X's java.ext.path by default. Worse, the old Java 3D 1.3 crashes
Oracle Java 8.

I thought of two different ways to address this new difficulty:

1. Change the package prefix to avoid a clash.
2. Manually manipulate the java.ext.path.

Option 2 is a deployment-centric solution, which only solves the problem
in environments where tweaking the Java runtime parameters is possible.
It is also a pain because there is no way to simply blacklist the
undesirable directory; instead, the java.ext.path must be built from the
ground up to include all the directories which Java includes by default,
except for /System/Java/Library/Extensions. Furthermore, there _are_
libraries in said directory that might continue to be useful for ImageJ
users, such as the support for AppleScript. It would be a shame to lose
access to _all_ Apple-specific Java extensions, just because of three
bad apples (the old Java 3D 1.3 JARs).

Therefore, this commit opts for option 1.